+++ src/macintosh.h
@@ -10,12 +10,13 @@
 
 char        *iokit_info(char *);
 
+static void execute_command(char const *const, char *const);
+
 void        get_ram_usage(char *),
             get_kernel(char *),
             get_complete_os(char *),
             get_resolution(char *),
             get_gpu(char *),
-            execute_command(char const *const, char *const),
             get_battery_procentage(char *);
 
 uint64_t    get_mem_from_vm_stat();